Improvements
 - move `getFilterFromLabel` and `getLabelFromFilter` in the SDK
 - improve `addLabelFilter` and `removeLabelFilter` implementation
 - improve searchInput code readability
