+++ libs/sdk-core/src/lib/db/search/index.ts
@@ -1,2 +1,3 @@
+export * from './filter';
 export * from './search.models';
 export * from './search';
